+++ example/common/config.ts
@@ -32,7 +32,9 @@ export default {
       'segment-picker',
       'textarea',
       'float-ball',
-      'input'
+      'input',
+      'tab-bar',
+      'swipe'
     ],
     introduce: [
       {
@@ -44,7 +46,9 @@ export default {
           'divider',
           'float-ball',
           'loading',
-          'collapse'
+          'collapse',
+          'tab-bar',
+          'swipe'
         ]
       },
       {
@@ -61,7 +65,8 @@ export default {
           'date-picker',
           'time-picker',
           'rate',
-          'switch'
+          'switch',
+          'slider'
         ]
       },
       {