+++ packages/mpx-cube-ui/lib/components/button/loading/loading.js
@@ -1,2 +1,5 @@
-import { createComponent } from '../../../common/helper/create-component';
-createComponent({});
+import { createComponent } from '@mpxjs/core';
+import rnMixin from './rn-mixin';
+createComponent({
+    mixins: [rnMixin]
+});