+++ packages/mpx-cube-ui/lib/components/popup/popup.js
@@ -44,9 +44,16 @@ createComponent({
             type: Object,
             value: {}
         },
+        // 是否移除catchtouch事件
         removeCatchTouch: {
             type: Boolean,
             value: false
+        },
+        // @dosHide
+        // 最外层view 的 pointerEvents配置
+        pointerEvents: {
+            type: String,
+            value: ''
         }
     },
     data: {
@@ -55,6 +62,21 @@ createComponent({
         display: false
     },
     computed: {
+        popupStyle() {
+            const style = { zIndex: this.zIndex };
+            if (this.pointerEvents) {
+                style.pointerEvents = this.pointerEvents;
+            }
+            if (!this.isVisible) {
+                if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
+                    style.pointerEvents = 'box-none';
+                }
+                else {
+                    style.pointerEvents = 'none';
+                }
+            }
+            return style;
+        },
         rootClass() {
             const cls = {
                 'cube-popup_mask': this.mask,
@@ -72,23 +94,29 @@ createComponent({
             }
             // eslint-disable-next-line
             // @ts-ignore
-            if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
+            if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
                 cls[`cube-popup-${this.display ? 'show' : 'hide'}`] = true;
             }
             return cls;
         },
-        maskOpacity() {
-            // eslint-disable-next-line
-            // @ts-ignore
+        maskStyle() {
+            const style = {};
             if (__mpx_mode__ !== 'ios' &&
                 __mpx_mode__ !== 'android' &&
+                __mpx_mode__ !== 'harmony' &&
                 this.styleConfig.mask?.visibleOpacity &&
                 this.isVisible) {
-                return {
-                    opacity: this.styleConfig.mask.visibleOpacity
-                };
+                style.opacity = this.styleConfig.mask.visibleOpacity;
             }
-            return {};
+            if (!this.isVisible) {
+                if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
+                    style.pointerEvents = 'box-none';
+                }
+                else {
+                    style.pointerEvents = 'none';
+                }
+            }
+            return style;
         }
     },
     methods: {