+++ packages/mpx-cube-ui/lib/components/popup/rn-mixin.js
@@ -7,14 +7,15 @@ let mixin = {
 };
 // eslint-disable-next-line
 // @ts-ignore
-if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
+if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
     mixin = {
         data: {
             animationData: {},
             rootAnimationData: {},
             maskAnimationData: {},
             ANIMATION_PRESET: {},
-            contentRect: {}
+            contentRect: {},
+            contentTranslateStyle: {}
         },
         computed: {
             contentInfo() {
@@ -44,8 +45,11 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
                     },
                     'cube-popup_transition': (animationOptions) => {
                         if (!this.isVisible) {
-                            setTimeout(() => {
-                                this.display = false;
+                            clearTimeout(this.dispalyTimer);
+                            this.dispalyTimer = setTimeout(() => {
+                                if (!this.isVisible) {
+                                    this.display = false;
+                                }
                                 // fix 玄学，不加 100ms ，drn 动画会非常卡
                             }, animationOptions.duration + 100);
                         }
@@ -77,6 +81,7 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
                 this.$watch('isVisible', (n, o) => {
                     if (!!n === !!o)
                         return;
+                    this.transitionendTimer && clearTimeout(this.transitionendTimer);
                     this.rnAnimation({
                         duration: 300,
                         timingFunction: 'ease-out'
@@ -93,16 +98,26 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
             translateAnimation(animationOptions, axis, start) {
                 const hasTranslate = !!this.animation;
                 const animation = this.animation || (this.animation = mpx.createAnimation(animationOptions));
+                this.targetTranslate = `translate${axis}`;
                 if (this.isVisible) {
                     if (hasTranslate) {
-                        animation[`translate${axis}`](start).step({ duration: 0 });
+                        animation[this.targetTranslate](start).step({ duration: 0 });
                     }
-                    animation[`translate${axis}`](0).step();
+                    animation[this.targetTranslate](0).step();
                 }
                 else {
-                    animation[`translate${axis}`](start).step();
+                    animation[this.targetTranslate](start).step();
                 }
                 this.animationData = animation.export();
+                this.transitionendTimer = setTimeout(() => {
+                    this.transitionend();
+                }, animationOptions.duration);
+            },
+            transitionend() {
+                if (this.isVisible && this.targetTranslate) {
+                    // 触发重新渲染
+                    this.contentTranslateStyle = {};
+                }
             },
             // @vuese
             // 仅 rn 使用，当内容元素高度变化后调用。用于更新动画高度
@@ -110,7 +125,14 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
                 return new Promise((resolve) => {
                     this.$refs['popup-content'].boundingClientRect((res) => {
                         if (res) {
-                            this.contentRect = res;
+                            const keys = Object.keys(res);
+                            const contentRect = this.contentRect;
+                            for (const key in keys) {
+                                if (res[key] !== contentRect[key]) {
+                                    this.contentRect = res;
+                                    break;
+                                }
+                            }
                         }
                         else {
                             this.boundingClientRectFaill = this.boundingClientRectFaill || 0;