+++ packages/mpx-cube-ui/src/components/dialog/dialog.ts
@@ -52,7 +52,7 @@ createComponent({
     },
     /**
      * @description 图标类型（自动添加`cubeic-`前缀）
-     * @optional 图标 Icon，更多选择参见[内置 Icon](https://www.mpxjs.cn/mpx-cube-ui/demo-theme-default/index.html#/pages/icon/index)
+     * @optional 图标 Icon，更多选择参见[内置 Icon](https://www.mpxjs.cn/mpx-cube-ui/example/index.html#/pages/icon/index)
      */
     icon: {
       type: String,
@@ -111,9 +111,17 @@ createComponent({
     },
     /**
      * @description 通过 wx:style透传样式, 里面的每项分别修改对应位置的样式
-     * @optional styleConfig = { headIcon: '' }
+     * @optional styleConfig = { headIcon, main, close  }
      */
     styleConfig: {
+      type: Object,
+      value: {}
+    },
+    /**
+     * @description 通过 wx:style 透传样式给popup, 里面的每项分别修改对应位置的样式
+     * @optional styleConfig = { content: '' }
+     */
+    popupStyleConfig: {
       type: Object,
       value: {}
     }
@@ -135,7 +143,7 @@ createComponent({
       return this.type === 'prompt'
     },
     containerClass () {
-      return `cube-dialog-${this.type}`
+      return `cube-dialog-container cube-dialog-${this.type}`
     },
     btnsClass () {
       return {
@@ -145,7 +153,8 @@ createComponent({
     cancelBtnClass () {
       return {
         [`${btnClsPrefix}-highlight`]: !!this._cancelBtn.active,
-        [`${btnClsPrefix}_disabled`]: !!this._cancelBtn.disabled
+        [`${btnClsPrefix}_disabled`]: !!this._cancelBtn.disabled,
+        [`${btnsClsPrefix}-cancel-border-right-1px`]: this.isConfirm || this.isPrompt
       }
     },
     confirmBtnClass () {