+++ packages/mpx-cube-ui/src/components/switch/rn-mixin.ts
@@ -3,7 +3,7 @@ import mpx, { getMixin } from '@mpxjs/core'
 let mixin = {} as Parameters<typeof getMixin>[0]
 // eslint-disable-next-line
 // @ts-ignore
-if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
+if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
   type ANIMATION_PRESET = Record<
     string,
     (
@@ -20,22 +20,22 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
       created() {
         this.ANIMATION_PRESET = {
           'cube-switch-BGC': (animationOptions) => {
-            const animation = this.maskAnimation || (this.maskAnimation = mpx.createAnimation({ ...animationOptions }))
+            const animation = this.bgAnimation || (this.bgAnimation = mpx.createAnimation({ ...animationOptions }))
             if (this.isOn) {
-              const onBGC = this.switchOnGBC || '#FF6435'
-              animation.backgroundColor(onBGC).step()
+              const onBGC = this.switchOnBGC || this.switchOnGBC || '#FF6435'
+              animation.backgroundColor(onBGC).step({ duration: animationOptions.duration })
             } else {
-              const defaultBGC = this.switchDefaultGBC || '#EAEAEA'
-              animation.backgroundColor(defaultBGC).step()
+              const defaultBGC = this.switchDefaultBGC || this.switchDefaultGBC || '#EAEAEA'
+              animation.backgroundColor(defaultBGC).step({ duration: animationOptions.duration })
             }
             this.switchAnimationData = animation.export()
           },
           'cube-switch-left': (animationOptions) => {
-            const animation = this.maskAnimation || (this.maskAnimation = mpx.createAnimation({ ...animationOptions }))
+            const animation = this.handleAnimation || (this.handleAnimation = mpx.createAnimation({ ...animationOptions }))
             if (this.isOn) {
-              animation.left(18).step()
+              animation.left(18).step({ duration: animationOptions.duration })
             } else {
-              animation.left(2).step()
+              animation.left(2).step({ duration: animationOptions.duration })
             }
             this.switchHandleAnimationData = animation.export()
           }
@@ -48,9 +48,14 @@ if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
               duration: 300,
               timingFunction: 'ease'
             })
-          },
-          { immediate: true }
+          }
         )
+        if (this.isOn) {
+          this.rnAnimation({
+            duration: 0,
+            timingFunction: 'ease'
+          })
+        }
       }
     },
     methods: {