+++ packages/mpx-cube-ui/src/components/toast/toast.ts
@@ -6,7 +6,7 @@ createComponent({
   properties: {
     /**
      * @description 图标类型（自动添加`cubeic-`前缀）
-     * @optional 图标 Icon，更多选择参见[内置 Icon](https://www.mpxjs.cn/mpx-cube-ui/demo-theme-default/index.html#/pages/icon/index)
+     * @optional 图标 Icon，更多选择参见[内置 Icon](https://www.mpxjs.cn/mpx-cube-ui/example/index.html#/pages/icon/index)
      */
     icon: {
       type: String,
@@ -36,7 +36,16 @@ createComponent({
     /**
      * @description 提示信息文案（一行最多只能展示十二个文字最多展示两行）
      */
-    txt: String
+    txt: String,
+    // 未配置mask时，toast外部区域是否可以点击
+    preventOutside: {
+      type: Boolean,
+      value: false
+    },
+    transition: {
+      type: String,
+      value: 'fade'
+    }
   },
   lifetimes: {
     ready() {
@@ -48,10 +57,18 @@ createComponent({
     timer: null
   },
   computed: {
+    pointerEvents() {
+      if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
+        if (this.preventOutside) return 'box-only'
+        if (!this.mask) return 'none'
+      }
+      if (this.preventOutside) return 'initial'
+      return ''
+    },
     tostTipClass() {
       // eslint-disable-next-line
       // @ts-ignore
-      if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android') {
+      if (__mpx_mode__ === 'ios' || __mpx_mode__ === 'android' || __mpx_mode__ === 'harmony') {
         return {
           'cube-toast-tip-icon': !!this.icon
         }