+++ scripts/jest-mpx-resolver/tab-bar.js
@@ -1,0 +1,1 @@
+export default '../tab-bar/index.mpx'