+++ scripts/jest-mpx-resolver/tab.js
@@ -1,0 +1,1 @@
+export default '../tab-bar/tab.mpx'