+++ docs/guide/changelog.md
@@ -1,9 +1,134 @@
-## 更新日志
+# Changelog
 
-<card>
+## [1.4.14] - 2026-03-31
+- slider 组件，优化 rn 下交互
 
-### 1.0.0（2024.03.13）
-#### mpx-cube-ui 项目开源
-- Initial commit([11019fb](https://github.com/didi/mpx-cube-ui/commit/11019fbfdece8ea0e76258fb11bd8fb557c1a190))
+## [1.4.13] - 2026-03-31
+- checkbox-single 组件，支持 rn 
 
-</card>
+## [1.4.12] - 2026-03-13
+- 新增 slider 组件，支持三端
+
+## [1.4.11] - 2026-02-03
+- popup ali fix 滚动穿透问题
+
+## [1.4.10] - 2026-02-03
+- popup rn fix 动画时调用 initContentRect，导致动画卡顿
+
+## [1.4.9] - 2026-01-29
+- toast 增加动画配置，可关闭 toast 显、隐动画
+
+## [1.4.8] - 2026-01-20
+- popup rn 增加 zIndex 配置
+
+## [1.4.7] - 2026-01-13
+- icon、button 组件 css 编译问题修复（React Native）
+
+## [1.4.6] - 2026-01-08
+- toast 组件支持配置 preventOutside：未配置mask时，toast外部区域是否可以点击
+
+## [1.4.5] - 2025-12-05
+- dialog 对齐小程序、web、RN 关闭按钮样式
+
+## [1.4.4] - 2025-12-05
+- button 支持阴影（React Native）
+
+## [1.4.3] - 2025-11-11
+- 同步1.3.15版本，dialog 增加 styleConfig 样式配置（React Native）
+- 同步1.3.14版本，popup 隐藏时不阻挡点击，show hide 时闪烁修复
+
+## [1.4.2] - 2025-11-10
+- 修复switch props命名错误
+
+## [1.4.1] - 2025-11-06
+- switch 组件新增 changeOnClick，控制点击时是否需要切换状态
+
+## [1.4.0] - 2025-11-06
+- RN 组件适配鸿蒙
+
+## [1.3.17] - 2026-01-15
+- popup ali fix 滚动穿透问题
+
+## [1.3.16] - 2025-12-05
+- button 支持阴影（React Native）
+- dialog 对齐小程序、web、RN 关闭按钮样式
+
+## [1.3.15] - 2025-11-11
+- dialog 增加 styleConfig 样式配置（React Native）
+
+## [1.3.14] - 2025-11-11
+- popup 隐藏时不阻挡点击，show hide 时闪烁修复
+
+## [1.3.13] - 2025-11-10
+- switch 组件新增 style 配置
+- switch 支持 changeOnClick
+- 修复switch props命名错误
+
+## [1.3.12] - 2025-10-30
+
+### Added
+- tip 支持 rn
+
+## [1.3.11] - 2025-10-13
+
+### Fixed
+- rn popup 隐藏动画丢失
+
+## [1.3.10] - 2025-09-22
+
+### Fixed
+- 对齐小程序、web，修复dialog组件按钮按钮之间没有分割线（React Native）
+
+
+## [1.3.9] - 2025-09-18
+
+### Added
+- dialog组件添加close按钮 (React Native)
+
+## [1.3.8] - 2025-09-12
+
+### Fixed
+- 修复popup组件中包含input时input无法点击的问题 (React Native)
+
+## [1.3.7] - 2025-08-27
+
+### Fixed
+- 修复mpx-cube-ui首页在skyline引擎下无法滑动的问题
+- tip组件对于skyline引擎的适配
+- 修改箭头CSS样式类命名
+
+## [1.3.6] - 2025-08-14
+
+### Fixed
+- dialog (小程序、web) 内容区删除未使用的css透明度属性，修复skyline 时动画问题
+
+## [1.3.5] - 2025-08-12
+
+### Fixed
+- 修复 tip （小程序、web）animationend报错问题
+
+## [1.3.4] - 2025-08-11
+
+### Fixed
+- 修复 popup （小程序、web） 添加 positive: relative 兼容 iphone 13 展示问题
+
+## [1.3.3] - 2025-08-02
+
+### Fixed
+- 修复 rn  switch 默认选中时动画问题
+
+## [1.3.2] - 2025-07-31
+
+### Fixed
+
+- 修复 React Native toast mask false 不阻塞点击
+
+## [1.3.1] - 2025-07-XX
+
+### Fixed
+- 修复 React Native dialog container默认继承宽度
+
+## [1.3.0] - 2025-06-XX
+
+### Added
+- 新增input组件 (React Native )